backend/requirements.txt
flask
python-dotenv
supabase
pandas
statsmodels
flask-cors